resolve flash attention tests

+++ csrc/src/lib.rs
@@ -643,7 +643,7 @@ impl FlashAttentionVarLen {
             let (block_table_storage, block_table_layout) = block_table.storage_and_layout();
             let block_table_ptr = match &*block_table_storage {
                 candle_core::Storage::Cuda(c) => {
-                    let cuda_slice = c.as_cuda_slice::<i64>()?;
+                    let cuda_slice = c.as_cuda_slice::<u32>()?;
                     let block_table = cuda_slice.slice(block_table_layout.start_offset()..);
                     let block_table_stride = block_table_layout.stride();
                     let block_table_rank = block_table_stride.len();
